feat: Add orientation adjustment in Controller class
This commit adds a new feature to the Controller class in the mpc.py file. It introduces an orientation adjustment to ensure that the state's orientation is consistent with the orientation trajectory. This adjustment is made by checking the sign of the state's orientation and comparing it with the sign of the orientation at the end of the trajectory. If they differ, the state's orientation is incremented by 2*pi.

Refactor: Update max_angular_vel in Robot class

This commit refactors the max_angular_vel parameter in the Robot class in the robot.py file. The value is changed from 1 to 3.14 to provide a more realistic maximum angular velocity for the robot.

fix: Update orientation_solver in math_utils.py

This commit fixes a bug in the orientation_solver function in the math_utils.py file. When the orientations have different signs and one of them is greater than 3*pi/4 or less than -3*pi/4, the function now correctly adjusts the second orientation by adding 2*pi.